pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "compliance-checker"
version = "0.1.0"
description = "AI-powered dataset usage compliance checker"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]
